+++ moview/modules/answer_evaluator/answer_analyzer.py
@@ -4,7 +4,7 @@ from langchain.prompts.chat import (
     ChatPromptTemplate,
     HumanMessagePromptTemplate
 )
-from moview.modules.prompt_loader.prompt_loader import PromptLoader
+from moview.utils.prompt_loader import PromptLoader
 from moview.environment.llm_factory import LLMModelFactory
 from moview.loggers.mongo_logger import *
