--- moview/modules/prompt_loader/__init__.py
@@ -1,1 +1,0 @@
-from moview.modules.prompt_loader.prompt_loader import PromptLoader
