+++ moview/utils/prompt_loader/__init__.py
@@ -1,0 +1,1 @@
+from moview.utils.prompt_loader.prompt_loader import PromptLoader
